#19 수요일 FollowUp - 1 파티클 출력 안되지만 지오메트리 셰이더까지 가볼 예정

+++ Inho/InhoDx2D/Project/Engine/particle.fx
@@ -19,7 +19,7 @@ struct VS_OUT
 {
     float4 vPosition : SV_Position;
     float2 vUV : TEXCOORD;
-    float InstID : FOG;
+    uint  InstID : FOG;
 };
 
 VS_OUT VS_Particle(VS_IN _in)
